// app/components/status-content/types.ts
import { LucideIcon } from "lucide-react";

export type ProjectProps = {
    Logo: LucideIcon;
    title: string;
    description: string;
    nomUrl: string;
    url: string;
};

export type FormationProps = {
    Logo: LucideIcon;
    title: string;
    description: string;
    date: string;
    url: string;
};